+++ .jshintrc
@@ -1,30 +1,25 @@
 {
-    "browser": true,
-    "bitwise": true,
-    "camelcase": true,
-    "curly": false,
-    "eqeqeq": true,
-    "immed": true,
-    "indent": 2,
-    "latedef": true,
-    "newcap": true,
-    "noarg": true,
-    "quotmark": "double",
-    "regexp": true,
-    "undef": true,
-    "unused": true,
-    "globals": {
-      "requirejs": false,
-      "require": false,
-      "define": false,
-      "console": false,
-      "jQuery": false,
-      "$": false,
-      "_gaq": false
-    },
-    "strict": true,
-    "trailing": true,
-    "smarttabs": true,
-    "expr": true
+  "esnext": true,
+  "browser": true,
+  "bitwise": true,
+  "camelcase": true,
+  "curly": false,
+  "eqeqeq": true,
+  "immed": true,
+  "indent": 2,
+  "latedef": true,
+  "newcap": true,
+  "noarg": true,
+  "quotmark": "double",
+  "regexp": true,
+  "undef": true,
+  "unused": true,
+  "globals": {
+    "console": false,
+    "jQuery": false,
+    "_gaq": false
+  },
+  "trailing": true,
+  "smarttabs": true,
+  "expr": true
 }
-
