+++ .npmignore
