--- .travis.yml
@@ -1,8 +1,0 @@
-language: node_js
-node_js:
-- "0.10"
-before_script:
-- npm install grunt-cli -g
-- npm install
-script:
-- grunt travis
