+++ example/index.html
@@ -6,8 +6,8 @@
 
     <script src="../dist/modernizr.js"></script>
 
-    <link rel="stylesheet" href="../bower_components/neue/dist/neue.js"/>
-    <link rel="stylesheet" href="../bower_components/neue/dist/neue.css"/>
+    <link rel="stylesheet" href="../node_modules/neue/dist/neue.js"/>
+    <link rel="stylesheet" href="../node_modules/neue/dist/neue.css"/>
     <link rel="stylesheet" href="../dist/modal.css"/>
   </head>
   <body>
@@ -20,9 +20,9 @@
           </div>
         </div>
 
-        <header role="banner">
+        <header class="header" role="banner">
           <div class="wrapper">
-            <h1 class="__title">Modal Example</h1>
+            <h1 class="header__title">Modal Example</h1>
           </div>
         </header>
 
@@ -39,14 +39,15 @@
       </div>
 
       <div data-modal id="test-modal">
-        <p>It's working!</p>
+        <div class="modal__block">
+          <p>It's working!</p>
+        </div>
       </div>
     </div>
 
 
-    <script src="../bower_components/jquery/jquery.js"></script>
+    <script src="../node_modules/jquery/dist/jquery.js"></script>
     <script src="../dist/modal.js"></script>
-
   </body>
 </html>
