+++ package.json
@@ -7,28 +7,39 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/DoSomething/neue.git"
+    "url": "https://github.com/DoSomething/modal.git"
   },
   "scripts": {
     "test": "grunt test"
   },
+  "dependencies": {
+    "jquery": "^1.11.0"
+  },
   "devDependencies": {
     "autoprefixer-core": "^4.0.2",
+    "babel-core": "^4.7.0",
+    "babel-loader": "^4.2.0",
     "css-mqpacker": "^2.0.0",
-    "each-async": "^1.1.1",
+    "grunt": "^0.4.0",
     "grunt-bump": "^0.0.14",
     "grunt-cli": "^0.1.13",
     "grunt-contrib-clean": "^0.6.0",
     "grunt-contrib-copy": "^0.7.0",
-    "grunt-contrib-jshint": "^0.7.1",
-    "grunt-contrib-qunit": "^0.3.0",
+    "grunt-contrib-jshint": "^0.11.1",
+    "grunt-contrib-qunit": "^0.5.2",
     "grunt-contrib-requirejs": "^0.4.0",
+    "grunt-contrib-uglify": "^0.8.0",
     "grunt-contrib-watch": "^0.6.1",
     "grunt-modernizr": "^0.6.0",
     "grunt-postcss": "^0.2.0",
+    "grunt-sass": "^0.18.0",
+    "grunt-webpack": "^1.0.8",
     "jshint-stylish": "^1.0.0",
-    "load-grunt-config": "^0.16.0",
-    "node-sass": "^2.0.0-beta",
-    "time-grunt": "^1.0.0"
+    "load-grunt-tasks": "^3.1.0",
+    "neue": "DoSomething/neue#v6.3.0",
+    "qunitjs": "^1.17.1",
+    "time-grunt": "^1.0.0",
+    "webpack": "^1.4.5",
+    "webpack-dev-server": "^1.0.0"
   }
 }
