--- scss/modal-build.scss
@@ -1,28 +1,0 @@
-//                 ____  _____    _   __
-//                / __ \/ ___/   / | / /__  __  _____
-//               / / / /\__ \   /  |/ / _ \/ / / / _ \
-//              / /_/ /___/ /  / /|  /  __/ /_/ /  __/
-//             /_____//____/  /_/ |_/\___/\__,_/\___/
-//
-//  ======================== Buildfile ========================
-//
-//  This file builds `neue.css`. You can also use it as an
-//  example for including Neue in another app's build.
-//
-
-// This is the path where Neue's assets can be found. Be
-// sure to include the trailing slash!
-// ex: $neue-path: "bower_components/neue/assets/"
-$neue-path: "assets/";
-
-// The $asset-path variable can be used by your client-app
-// to simplify including asset URLs.
-$asset-path: "";
-
-// Variables
-$tablet: "(min-width: 760px)";
-$base-spacing: 24px;
-$light-gray: #ddd;
-
-// Now that we have everything in order, let's build Neue!
-@import "modal";
