--- scss/styleguide.scss
@@ -1,307 +1,0 @@
-//                 ____  _____    _   __
-//                / __ \/ ___/   / | / /__  __  _____
-//               / / / /\__ \   /  |/ / _ \/ / / / _ \
-//              / /_/ /___/ /  / /|  /  __/ /_/ /  __/
-//             /_____//____/  /_/ |_/\___/\__,_/\___/
-//
-//   ==================== Styleguide Styles ====================
-//
-//   These are styles that are only used in the styleguide, and
-//   should not be available in client applications.
-
-@import "../bower_components/susy/sass/susy";
-
-@import "_utilities/functions";
-@import "_utilities/variables";
-@import "_utilities/mixins";
-
-// FixedSticky
-.fixedsticky-dummy {
-  display: none;
-}
-
-.fixedsticky-on + .fixedsticky-dummy {
-  display: block;
-}
-
-
-// Styleguide!
-main {
-  @include clearfix;
-}
-
-.styleguide-sidebar {
-  padding-top: $base-spacing;
-
-  @include media($tablet) {
-    @include span(4 rtl);
-    position: sticky;
-    top: 0;
-
-    // Fallback for position:sticky, through Filament Group's fixed-sticky
-    &.fixedsticky-on {
-      position: fixed;
-      width: 100%;
-      max-width: 1440px;
-      padding: 0;
-
-      .waypoints {
-        position: absolute;
-        width: span(4);
-        padding: $base-spacing gutter();
-        right: 0;
-        display: inline-block;
-      }
-    }
-
-    &.fixedsticky-off {
-      position: static;
-    }
-  }
-}
-
-.styleguide-content {
-  @include clearfix;
-  @include span(100%);
-  padding-top: $base-spacing;
-
-  @include media($tablet) {
-    @include span(10);
-    @include push(2);
-  }
-
-  // Override galleries in styleguide content to fit to grid.
-  > .gallery {
-    margin: ($base-spacing / 2) (-$base-spacing / 2);
-  }
-}
-
-.styleguide-example {
-  margin: 15px 0 50px;
-  background: rgba(255, 255, 255, 0.5);
-}
-
-.styleguide-example-content {
-  border-radius: 2px;
-  border: 1px solid #ddd;
-}
-
-.styleguide-example > h3 {
-  margin: 0;
-  padding: 5px 0;
-  color: #333;
-  font-size: 12px;
-  text-transform: uppercase;
-}
-
-.styleguide-example > h3 a {
-  color: #333;
-}
-
-.styleguide-example > h3 em {
-  float: right;
-  text-transform: none;
-  font-style: normal;
-  font-weight: normal;
-  color: #999;
-}
-
-.styleguide-example .styleguide-description {
-  padding: 10px 15px;
-  background: #f1f1f1;
-  border-bottom: 1px solid #ddd;
-}
-
-.styleguide-example .styleguide-description p:first-child {
-  margin-top: 0;
-}
-
-.styleguide-example .styleguide-description p:last-child {
-  margin-bottom: 0;
-}
-
-.styleguide-example .styleguide-element {
-  overflow: hidden;
-  position: relative;
-  padding: 20px;
-  min-height: 100px;
-}
-
-.styleguide-example .styleguide-element + .styleguide-element {
-  padding-top: 15px;
-  border-top: 1px solid #eee;
-}
-
-.styleguide-example .styleguide-element .styleguide-modifier-name {
-  display: block;
-  position: absolute;
-  top: 0;
-  right: 0;
-  padding: 5px 8px;
-  font-size: 11px;
-  color: #999;
-  background: #f9f9f9;
-  border: 1px solid #eee;
-  border-top: 0;
-  z-index: 10;
-}
-
-.styleguide-example .styleguide-html {
-  padding: 10px 25px 10px 5px;
-  background: #efefef;
-  border-top: 1px solid #dde7ea;
-  font-size: 18px;
-  overflow: auto;
-}
-
-.hljs {
-  background: transparent !important;
-  display: inline !important;
-  padding-right: 25px !important;
-}
-
-.styleguide-example .styleguide-html .highlight {
-  background: none;
-}
-
-.styleguide-example ul.styleguide-modifiers {
-  margin: 0 0 0 10px;
-}
-
-.styleguide-example ul.styleguide-modifiers li {
-  list-style-type: none;
-  margin-left: 0;
-}
-
-.styleguide-example ul.styleguide-modifiers li strong {
-  font-family: Monaco, monospace;
-  font-size: 12px;
-  font-weight: normal;
-  color: #222;
-}
-
-.styleguide-modifier {
-  list-style-type: none;
-  padding: 0.5em 1em;
-  margin: 0;
-}
-
-.styleguide-modifier li + li {
-  margin-top: 0;
-}
-
-.kss-dark .styleguide-element {
-  background: #666;
-}
-
-.kss-min-height .styleguide-element {
-  min-height: 300px;
-}
-
-.font-tile {
-  font-size: 27px;
-  margin: 18px 0;
-  padding: 18px;
-  border-radius: 2px;
-  background: #f1f1f1;
-  border: 1px solid #ddd;
-}
-
-.font-tile code {
-  display: block;
-  font-size: 12px;
-  color: #999;
-  font-weight: 400;
-  line-height: 1.5;
-  padding-top: 5px;
-}
-
-.color-tile {
-  width: 100%;
-  background: #eee;
-  color: #fff;
-  text-shadow: $text-shadow;
-  padding: ($base-spacing * 2) $base-spacing;
-  border-radius: 2px;
-
-  & code {
-    color: rgba(255, 255, 255, 0.7);
-  }
-}
-
-.color-tile__value {
-  display: block;
-}
-
-.styleguide-tip {
-  &:before {
-    @include icomoon-icon;
-    content: "\f000 ";
-    position: relative;
-    top: 1px;
-    padding-right: 4px;
-  }
-}
-
-.styleguide-footer {
-  margin-top: $section-spacing;
-  margin-bottom: $section-spacing;
-
-  p {
-    color: #aaa;
-    font-size: $font-small;
-  }
-
-  a {
-    color: #666;
-  }
-}
-
-header[role="banner"].-styleguide {
-  background: url("../styleguide/assets/example-poster.jpg") no-repeat;
-  background-size: cover;
-}
-
-.semver-badge {
-  font-size: 16px;
-  vertical-align: 50%;
-  background: rgba(#000, 0.25);
-  padding: 4px 8px;
-  text-shadow: none;
-  border-radius: 4px;
-}
-
-.intro-neue {
-  font-weight: 400;
-}
-
-.kss-markuplink {
-  color: $med-gray;
-  font-size: $font-smaller;
-  margin: 4px;
-
-  a {
-    color: $med-gray;
-
-    &:before {
-      @include icomoon-icon;
-      content: "\f05f ";
-      position: relative;
-      top: 1px;
-      padding-right: 6px;
-    }
-
-    &:hover {
-      text-decoration: none;
-    }
-  }
-}
-
-// Specific pattern overrides
-#message-callout--example---white,
-#carousel--example,
-#information-bar--example,
-#site-navigation--example--white {
-  background: #fff url("/styleguide/assets/pattern-dark-bg.png");
-}
-
