--- tasks/aliases.json
@@ -1,7 +1,0 @@
-{
-  "default": ["build", "lint", "test", "watch"],
-  "build": ["clean:dist", "sass", "postcss", "requirejs", "modernizr"],
-  "lint": ["jshint"],
-  "test": ["qunit"],
-  "travis": ["lint", "test"]
-}
