--- tasks/clean.js
@@ -1,3 +1,0 @@
-module.exports = {
-  "dist": ["dist/"]
-};
