--- tasks/jshint.js
@@ -1,11 +1,0 @@
-module.exports = {
-  "options": {
-    "jshintrc": true,
-    "reporter": require("jshint-stylish")
-  },
-  "all": [
-    "js/**/*.js",
-    "tests/**/*.js",
-    "!tests/lib/**/*.js"
-  ]
-};
