+++ wercker.yml
@@ -2,19 +2,10 @@ box: wercker/nodejs
 
 build:
   steps:
-    - script:
-        name: Install Bower components
-        code: |
-          export BOWER_STORAGE__CACHE="${WERCKER_CACHE_DIR}/wercker/bower"
-          mkdir -p $BOWER_STORAGE__CACHE
-          npm install bower
-          ./node_modules/bower/bin/bower install
     - npm-install
+    - grunt:
+        tasks: build
     - npm-test
-  after_steps:
-    - hipchat-notify:
-        token: $HIPCHAT_TOKEN
-        room-id: 255200
 
 deploy:
   steps:
@@ -27,8 +18,8 @@ deploy:
         name: Get package version
         code: export PACKAGE_VERSION="$(node -p 'var pkg = require("./package.json"); pkg.version;')"
     - script:
-        name: Use distribution ignore file
-        code: rm .gitignore && mv .distignore .gitignore
+        name: Use NPM ignore file for tag
+        code: rm .gitignore && mv .npmignore .gitignore
     - dfurnes/git-tag-version
   after_steps:
     - hipchat-notify:
